photoeditor/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0.1
djangorestframework==3.13.1
numpy==1.22.1
opencv-contrib-python==4.5.5.62
opencv-python==4.5.5.62
Pillow==9.0.0
pkg_resources==0.0.0
pytz==2021.3
sqlparse==0.4.2
